refactor(frontend): extract tabs from BookDedup + hooks from Library (FE-5, STR-4)
STR-4: BookDedup.tsx 2907→145 lines — extracted AIReviewTab (386 lines),
EmbeddingDedupTab (1441 lines, includes module-level cache + buildClusters
union-find utility), and AcousticDedupTab (996 lines, re-exports
AcousticBookMetadata + AcousticBookCard). Root component is now a pure tab router.

FE-5: Library.tsx 2018→1811 lines — extracted useLibraryQuery (229 lines,
book-fetch state + effects + auto-refresh + op-complete reload) and
useLibrarySelection (164 lines, selection state + 6 selection handlers).

TypeScript: 0 errors (npx tsc --noEmit clean from web/).

Part of audit remediation PR L.
